system-design/notification-service/requirements.txt
# Web framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# Kafka
kafka-python>=2.0.0,<3.0.0
confluent-kafka>=2.3.0,<3.0.0

# Database
asyncpg>=0.29.0,<0.30.0
databases[postgresql]>=0.8.0,<0.9.0
sqlalchemy>=2.0.0,<2.1.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Redis for rate limiting
redis>=5.0.0,<6.0.0

# HTTP requests for mock services
httpx>=0.25.0,<0.26.0
aiohttp>=3.9.0,<4.0.0

# Data validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Utilities
python-json-logger>=2.0.0,<3.0.0
structlog>=23.2.0,<24.0.0
tenacity>=8.2.0,<9.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-mock>=3.12.0,<4.0.0

# Development
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0 